+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- import `xlsx`: A detailled error message provides sheet name and file path, when a potentially empty sheet makes processing impossible.
+
 ## [0.54.0] - 2026-06-09
 
 ### Added